# 8277793: Support vector F2I and D2L cast operations for X86