--- a/src/hotspot/cpu/x86/c2_MacroAssembler_x86.cpp
+++ b/src/hotspot/cpu/x86/c2_MacroAssembler_x86.cpp
@@ -4070,8 +4070,8 @@ void C2_MacroAssembler::masked_op(int ideal_opc, int mask_len, KRegister dst,
  */
 
 void C2_MacroAssembler::vector_castD2L_evex(XMMRegister dst, XMMRegister src, XMMRegister xtmp1, XMMRegister xtmp2,
-                                            KRegister ktmp1, AddressLiteral double_sign_flip,
-                                            AddressLiteral max_long, Register scratch, int vec_enc) {
+                                            KRegister ktmp1, KRegister ktmp2, AddressLiteral double_sign_flip,
+                                            Register scratch, int vec_enc) {
   Label done;
   evcvttpd2qq(dst, src, vec_enc);
   evmovdqul(xtmp1, k0, double_sign_flip, true, vec_enc, scratch);
@@ -4080,43 +4080,44 @@ void C2_MacroAssembler::vector_castD2L_evex(XMMRegister dst, XMMRegister src, XM
   jccb(Assembler::equal, done);
 
   vpxor(xtmp2, xtmp2, xtmp2, vec_enc);
-  evcmppd(ktmp1, k0, src, src, Assembler::UNORD_Q, vec_enc);
-  evblendmpd(dst, ktmp1, dst, xtmp2, true, vec_enc);
+  evcmppd(ktmp2, k0, src, src, Assembler::UNORD_Q, vec_enc);
+  evmovdquq(dst, ktmp2, xtmp2, true, vec_enc);
 
-  evpcmpeqq(ktmp1, xtmp1, dst, vec_enc);
+  kxorwl(ktmp1, ktmp1, ktmp2);
   evcmppd(ktmp1, ktmp1, src, xtmp2, Assembler::NLT_US, vec_enc);
-  evmovdquq(xtmp1, max_long, vec_enc, scratch);
-  evblendmpd(dst, ktmp1, dst, xtmp1, true, vec_enc);
+  vpternlogq(xtmp2, 0x11, xtmp1, xtmp1, vec_enc);
+  evmovdquq(dst, ktmp1, xtmp2, true, vec_enc);
   bind(done);
 }
 
 void C2_MacroAssembler::vector_castF2I_avx(XMMRegister dst, XMMRegister src, XMMRegister xtmp1,
-                                           XMMRegister xtmp2, XMMRegister xtmp3, AddressLiteral float_sign_flip,
-                                           AddressLiteral max_int, Register scratch, int vec_enc) {
+                                           XMMRegister xtmp2, XMMRegister xtmp3, XMMRegister xtmp4,
+                                           AddressLiteral float_sign_flip, Register scratch, int vec_enc) {
   Label done;
   vcvttps2dq(dst, src, vec_enc);
   vmovdqu(xtmp1, float_sign_flip, scratch);
   vpcmpeqd(xtmp2, dst, xtmp1, vec_enc);
-  vpmovmskb(scratch, xtmp2, vec_enc);
-  testl(scratch, scratch);
+  vptest(xtmp2, xtmp2, vec_enc);
   jccb(Assembler::equal, done);
 
-  vpxor(xtmp2, xtmp2, xtmp2, vec_enc);
+  vpxor(xtmp4, xtmp4, xtmp4, vec_enc);
   vcmpps(xtmp3, src, src, Assembler::UNORD_Q, vec_enc);
-  vblendvps(dst, dst, xtmp2, xtmp3, vec_enc);
+  vblendvps(dst, dst, xtmp4, xtmp3, vec_enc);
 
-  vpcmpeqd(xtmp1, dst, xtmp1, vec_enc);
-  vpand(xtmp2, xtmp1, src, vec_enc);
-  vpxor(xtmp3, xtmp2, xtmp1, vec_enc);
-  vpand(xtmp3, xtmp3, xtmp1, vec_enc);
-  vmovdqu(xtmp1, max_int, scratch);
+  vpxor(xtmp2, xtmp2, xtmp3, vec_enc);
+  vpand(xtmp4, xtmp2, src, vec_enc);
+  vpxor(xtmp3, xtmp2, xtmp4, vec_enc);
+
+  vpcmpeqd(xtmp4, xtmp4, xtmp4, vec_enc);
+  vpxor(xtmp1, xtmp1, xtmp4, vec_enc);
+
   vblendvps(dst, dst, xtmp1, xtmp3, vec_enc);
   bind(done);
 }
 
 void C2_MacroAssembler::vector_castF2I_evex(XMMRegister dst, XMMRegister src, XMMRegister xtmp1, XMMRegister xtmp2,
-                                            KRegister ktmp1, AddressLiteral float_sign_flip,
-                                            AddressLiteral max_int, Register scratch, int vec_enc) {
+                                            KRegister ktmp1, KRegister ktmp2, AddressLiteral float_sign_flip,
+                                            Register scratch, int vec_enc) {
   Label done;
   vcvttps2dq(dst, src, vec_enc);
   evmovdqul(xtmp1, k0, float_sign_flip, false, vec_enc, scratch);
@@ -4125,13 +4126,13 @@ void C2_MacroAssembler::vector_castF2I_evex(XMMRegister dst, XMMRegister src, XM
   jccb(Assembler::equal, done);
 
   vpxor(xtmp2, xtmp2, xtmp2, vec_enc);
-  evcmpps(ktmp1, k0, src, src, Assembler::UNORD_Q, vec_enc);
-  evblendmps(dst, ktmp1, dst, xtmp2, true, vec_enc);
+  evcmpps(ktmp2, k0, src, src, Assembler::UNORD_Q, vec_enc);
+  evmovdqul(dst, ktmp2, xtmp2, true, vec_enc);
 
-  Assembler::evpcmpeqd(ktmp1, k0, xtmp1, dst, vec_enc);
+  kxorwl(ktmp1, ktmp1, ktmp2);
   evcmpps(ktmp1, ktmp1, src, xtmp2, Assembler::NLT_US, vec_enc);
-  evmovdquq(xtmp1, max_int, vec_enc, scratch);
-  evblendmps(dst, ktmp1, dst, xtmp1, true, vec_enc);
+  vpternlogd(xtmp2, 0x11, xtmp1, xtmp1, vec_enc);
+  evmovdqul(dst, ktmp1, xtmp2, true, vec_enc);
   bind(done);
 }