--- a/src/hotspot/cpu/x86/c2_MacroAssembler_x86.hpp
+++ b/src/hotspot/cpu/x86/c2_MacroAssembler_x86.hpp
@@ -290,14 +290,14 @@ public:
                  KRegister src1, KRegister src2);
 
   void vector_castF2I_avx(XMMRegister dst, XMMRegister src, XMMRegister xtmp1,
-                          XMMRegister xtmp2, XMMRegister xtmp3, AddressLiteral float_sign_flip,
-                          AddressLiteral max_int, Register scratch, int vec_enc);
+                          XMMRegister xtmp2, XMMRegister xtmp3, XMMRegister xtmp4,
+                          AddressLiteral float_sign_flip, Register scratch, int vec_enc);
 
   void vector_castF2I_evex(XMMRegister dst, XMMRegister src, XMMRegister xtmp1, XMMRegister xtmp2,
-                           KRegister ktmp1, AddressLiteral float_sign_flip,
-                           AddressLiteral max_int, Register scratch, int vec_enc);
+                           KRegister ktmp1, KRegister ktmp2, AddressLiteral float_sign_flip,
+                           Register scratch, int vec_enc);
 
   void vector_castD2L_evex(XMMRegister dst, XMMRegister src, XMMRegister xtmp1, XMMRegister xtmp2,
-                           KRegister ktmp1, AddressLiteral double_sign_flip,
-                           AddressLiteral max_long, Register scratch, int vec_enc);
+                           KRegister ktmp1, KRegister ktmp2, AddressLiteral double_sign_flip,
+                           Register scratch, int vec_enc);
 #endif // CPU_X86_C2_MACROASSEMBLER_X86_HPP