@jatin-bhateja 2021-11-26
> Is this change already covered by tests and micros?

Hi @neliasso , 

Yes we do have existing tests and micro which exercises these change.

MICRO : 
         test/micro/org/openjdk/bench/vm/compiler/TypeVectorOperations.java
TEST:  test/hotspot/jtreg/compiler/codegen/TestLongDoubleVect.java
          test/hotspot/jtreg/compiler/codegen/TestLongDoubleVect.java

BR
Jatin